refactor: Rebrand to pytest-iovis
It's become increasingly more clear that the core of this plugin isn't
the papermill integration, but instead the cosmetic changes done during
test collection.

Changing the name to iovis makes it more clear that this is meant
as a companion to Jupyter Notebooks, and isn't completely centered on
the relationship with papermill. Which goes in hand with the work
to make papermill an optional component.